+++ src/main/java/io/server/accountserviceserver/service/AccountService.java
@@ -2,8 +2,10 @@ package io.server.accountserviceserver.service;
 
 import io.server.accountserviceserver.entity.Account;
 
+import java.math.BigDecimal;
+
 public interface AccountService {
     Account getAccountById(Integer accountId);
 
-    Account updateAccountBalance(Integer accountId, Integer amount);
+    Account updateAccountBalance(Integer accountId, BigDecimal amount);
 }
